third_party/blink/web_tests/platform/generic/fast/dom/HTMLImageElement/image-picture-invalid-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS document.getElementById("canary").clientWidth is 75
PASS currentSrcFileName("canary") is "blue_rect.jpg?0"
PASS document.getElementById("div_end_picture").clientWidth is 800
PASS currentSrcFileName("div_end_picture") is "image-set-2x.png?2"
PASS document.getElementById("div_multiple_end").clientWidth is 800
PASS currentSrcFileName("div_multiple_end") is "image-set-2x.png?2"
PASS document.getElementById("source_in_div").clientWidth is 75
PASS currentSrcFileName("source_in_div") is "blue_rect.jpg"
PASS document.getElementById("not_in_picture").clientWidth is 75
PASS currentSrcFileName("not_in_picture") is "blue_rect.jpg"
